# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/complaints": {
            "get": {
                "description": "Get details of all complaints",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "complaints"
                ],
                "summary": "Get details of all complaints",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token header",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Complaint"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new complaint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "complaints"
                ],
                "summary": "Create complaint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token header",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Complaint"
                            }
                        }
                    }
                }
            }
        },
        "/complaints/user/{userId}": {
            "get": {
                "description": "Get details of complaint made by a specific user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "complaints"
                ],
                "summary": "Get all complaints of user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token header",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "userId",
                        "name": "leaseId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Complaint"
                            }
                        }
                    }
                }
            }
        },
        "/complaints/{complaintId}": {
            "get": {
                "description": "Get details of complaint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "complaints"
                ],
                "summary": "Get details of complaint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token header",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "complaintId",
                        "name": "leaseId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Complaint"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete the existing complaint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "complaints"
                ],
                "summary": "Delete complaint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token header",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Complaint"
                            }
                        }
                    }
                }
            }
        },
        "/leases": {
            "get": {
                "description": "Get details of all leases",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leases"
                ],
                "summary": "Get details of all leases",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token header",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Lease"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new lease",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leases"
                ],
                "summary": "Create lease",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token header",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Lease"
                            }
                        }
                    }
                }
            }
        },
        "/leases/{leaseId}": {
            "get": {
                "description": "Get details of leases",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leases"
                ],
                "summary": "Get details lease",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token header",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "leaseId",
                        "name": "leaseId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Lease"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update the existing lease",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leases"
                ],
                "summary": "Update lease",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token header",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Lease"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete the existing lease",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leases"
                ],
                "summary": "Delete lease",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token header",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Lease"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "models.Complaint": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "models.Lease": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "leaseEndDate": {
                    "type": "string"
                },
                "leaseStartDate": {
                    "type": "string"
                },
                "listingId": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
...
    }
}